Distributions settings finished

+++ eShop.Messaging/Models/Distribution/GetComissionAmountResponse.cs
@@ -3,9 +3,9 @@
     public class GetComissionAmountResponse : Messaging.Message, IResponse
     {
         public Guid AccountId { get; }
-        public decimal Amount { get; }
+        public double Amount { get; }
 
-        public GetComissionAmountResponse(Guid accountId, decimal amount)
+        public GetComissionAmountResponse(Guid accountId, double amount)
         {
             AccountId = accountId;
             Amount = amount;
