+++ eShop.Messaging/Models/Distribution/GetComissionSettingsResponse.cs
@@ -3,9 +3,9 @@
     public class GetComissionSettingsResponse : Messaging.Message, IResponse
     {
         public Guid AccountId { get; }
-        public decimal Amount { get; }
+        public double Amount { get; }
 
-        public GetComissionSettingsResponse(Guid accountId, decimal amount)
+        public GetComissionSettingsResponse(Guid accountId, double amount)
         {
             AccountId = accountId;
             Amount = amount;
